requirements.txt
fastapi
uvicorn[standard]
pydantic
firecrawl
python-dotenv
google-genai
